Rename GriptapeCloudKnowledgeBaseVectorStoreDriver to GriptapeCloudVectorStoreDriver
