+++ griptape/drivers/vector/griptape_cloud_vector_store_driver.py
